Table of Contents: Use a custom store subscription for observing headings
